Struct lex::LexerBuilder

pub struct LexerBuilder { /* fields omitted */ }

A helper struct for creating lexers in build scripts.

Instances of this stuct can be used in build scripts for creating lexers to be used in a package. It has two hidden paramaters that affect its behaviour: the name used by the builder when writing the finished lexer, and the location of the source file. These are private and can only be accessed or modified through the provided methods which automatically run correctness checks on the provided values.

Note that LexerBuilder’s are only intended to be run by Cargo as a part of a build script, and LexerBuilder::new will not return a builder unless the environment resembles that of a build script. See the official documentation for more details on build scripts.

Name and Location

The builder will write the lexer in the current builds OUT_DIR, with the name [name].rs. name must be a valid filename, with no seperators1 and no extension. If name is not set, but location is, name defaults to the filename for location minus the “.lex” extension. Otherwise it defaults to the name of the package with “_lex” added to the end.

The value for location must be an existing file ending with “.lex”. If location is not set then it defaults to name with “.lex” added to the end (if name and location are not set, name default to the package name plus “_lex”, so location would default to "[package name]_lex.lex`“.)

Environment

LexerBuilder’s rely on a number of environment variables to be set in order to understand the package being built and where to write files. These should be automatically set when the code is being run by Cargo as a part of a build script. For safety, LexerBuilder::new() will check the environment and will not return a builder unless it is properly configured.

// outside of build.rs

use lex::LexerBuilder;

assert!(LexerBuilder::new().is_none());

Reruns

Once the desired location for the source file of a lexer is set it is recommended to run set_rerun to set Cargo to monitor the source file for changes and rerun the build script when it detects one since the last build.

“/” on Unix/Redox and “\” on Windows. Note that, since this is running as part of a build script, it is the host system that matters, not the compilation target. This also means that you should be weary about using either / or \, regardless of the platform: something that works fine on your system may fail to compile when being built on a different platform! ↩

Implementations

impl LexerBuilder

pub fn new() -> Option<Self>

Creates a new instance of a LexerBuilder. Returns None if the environement is not configured properly. See here for more details.

Example

let mut builder = LexerBuilder::new().unwrap();
builder.set_default_values();

pub fn name(&self) -> Option<&OsStr>

Returns the name used by the builder if set. Returns None otherwise.

pub fn location(&self) -> Option<&Path>

Returns the location used by the builder. Returns None otherwise

pub fn with_name<S: Into<OsString>>(
&mut self,
name: S
) -> Result<&mut Self, LexerBuilderError>

Attempts to set the name for self to name.

If name is a valid filename (see here,) then this method returns a mutable reference to self, otherwise it returns an LexerBuilderError.

Example

use std::ffi::OsStr;
use lex::LexerBuilder;
let mut builder = LexerBuilder::new().unwrap();
let builder_name = builder
	.with_name("my_lexer")?
	.name();
assert_eq!(builder_name, Some(OsStr::new("my_lexer")));

pub fn at_location<S: Into<PathBuf>>(
&mut self,
location: S
) -> Result<&mut Self, LexerBuilderError>

Attempts to set the location for self to location.

If location is a valid path to a source file (see here,) then this method returns a mutable reference to self, otherwise it returns an LexerBuilderError.

Example

use std::path::Path;
use lex::LexerBuilder;

let mut builder = LexerBuilder::new().unwrap();
let builder_location = builder
	.at_location("lex-yacc/my_lexer.lex")?
	.location();

assert_eq!(builder_location, Some(Path::new("lex-yacc/my_lexer.lex")));

pub fn set_default_values(&mut self)

Sets the name and location to their default values, as described here. Calling this method directly is usually unecessary, as the fields will automatically be set to their default values as necessary when building a lexer. It can however be usefull for inspecting what the values would be.

Example

use std::ffi::OsStr;
use std::path::Path;
use lex::LexerBuilder;

let mut builder = LexerBuilder::new().unwrap();
builder.with_name("my_lexer").unwrap();

assert_eq!(builder.location(), None);

builder.set_default_values();

assert_eq!(builder.name(), Some(OsStr::new("my_lexer")));
assert_eq!(builder.location(), Some(Path::new("my_lexer.lex")));

pub fn set_rerun(&self) -> Option<()>

Tells Cargo to watch the source file for changes and rerun the build script if necessary. If no location is set, it returns None

Example

// in build.rs

use lex::{LexerBuilder, LexerBuilderError};

fn main() -> Result<(), LexerBuilderError> {
	let mut builder = LexerBuilder::new().unwrap();
	builder
		.with_name("my_lexer")
		.unwrap()
		.build()?
		.set_rerun().unwrap();
	Ok(())
}

pub fn output_file(&self) -> Option<PathBuf>

Returns the file path used when writing the lexer. If the name is not set, or if it cannot determin the out dir set by cargo, then it returns None.

Note that even if the name is set, the file may not exist if the builder hasn’t yet written a lexer there. This method just tells you where the builder would write to if you were to run build or something similar.

Example

use std::{env, path::PathBuf};
use lex::LexerBuilder;

let output_loc = LexerBuilder::new().unwrap()
	.with_name("my_lexer")?
	.output_file();
let path = PathBuf::from(
	format!(
		"{}/my_lexer.rs",
		env::var("OUT_DIR").unwrap()
	)
);
assert_eq!(output_loc, Some(path));

pub fn build(&mut self) -> Result<&mut Self, LexerBuilderError>

Builds the lexer, reading from the file at location and writing a file with the appropriate name to the out dir set by Cargo.

Example

// in build.rs

use lex::{LexerBuilder, LexerBuilderError};

fn main() -> Result<(), LexerBuilderError> {
	let mut builder = LexerBuilder::new().unwrap();
	builder
		.build()?
		.set_rerun().unwrap();
	Ok(())
}

pub fn build_from_str(
&mut self,
src: &str
) -> Result<&mut Self, LexerBuilderError>

Like build but reads its source from the provided src rather than than an input file.

Example

use lex::LexerBuilder;

let rules = r#"
	id : "[a-zA-Z_][0-9a-zA-Z_]*"
	ignore : "[[:space:]]+"
	int : "[0-9]+"
	error : "/""#;

let mut builder = LexerBuilder::new().unwrap();
builder.build_from_str(rules)?;

pub fn get_src(&mut self) -> Result<String, Error>

Reads the file at location and returns it as a string if successful. Otherwise returns an I/O error.