Trait lex::Lexer

pub trait Lexer<T> where
    for<'r, 's> &'r Self: IntoTokens<'s, T>, {
    fn new() -> Self;
}

A trait representing a lexer.

Lexer, IntoTokens, and Tokens

Tokens

Tokens (not to be confused with the struct Token) is implemented by types returned by lexers (particularly those generated by lex) when lexing a source. Types that implement Tokens<'s, T> iterate over the (lazily parsed) tokens of their source, returning a Result<Token<'s, T>, Error> (for Token and Error.) They also provide data on their location within the source for error reporting.

Lexer and IntoTokens

For technical reasons surrounding the limitations on associated types in traits, the interface for lexers is split into two traits: Lexer and IntoTokens. IntoTokens has a single method lex, which takes an input string as an argument and consumes self, returning a Tokens. A type which implements Lexer also implements IntoTokens on &'r Self for any lifetime 'r. This means that lex can be called directly on a Lexer as a borrowing method.

Example

use std::error::Error;
use lex::{IntoTokens, Lexer, Tokens};

type Result<T> = std::result::Result<T, Box<dyn Error>>;

enum TokenKind {
	Int,
	Plus,
	Minus,
}
use TokenKind::*;

fn calculate<L>(inp: &str) -> Result<i32>
where
	L: Lexer<TokenKind>,
	for<'r, 's> &'r L: IntoTokens<'s, TokenKind>,	// required by a bound in Lexer
{
	let lexer = L::new();

	let mut tokens = lexer.lex(inp);

	parse(&mut tokens)
}

fn parse<'s, T: Tokens<'s, TokenKind>>(tokens: &mut T) -> Result<i32> {
	let mut acc: i32 = match tokens.next() {
		Some(res) => res?.val().parse()?,
		None => return Err("empty input".into()),
	};
	loop {
		let op = match tokens.next() {
			Some(res) => res?,
			None => return Ok(acc),
		};
		acc += match op.kind() {
			Plus => expect_int(tokens)?,
			Minus => -expect_int(tokens)?,
			_ => return Err("missing operand".into()),
		};
	};
}

fn expect_int<'s, T: Tokens<'s, TokenKind>>(tokens: &mut T) -> Result<i32> {
	let err_mssg = || -> Box<dyn Error> { "missing integer".into() };
	let tok = tokens.next().ok_or_else(err_mssg)??;
	match tok.kind() {
		Int => Ok(tok.val().parse()?),
		_ => Err(err_mssg()),
	}
}

Required methods

fn new() -> Self